+++ lib/koyo.dart
@@ -2,5 +2,6 @@ library koyo;
 
 export 'src/core.dart';
 export 'src/extensions/ky_extensions.dart';
+export 'src/models.dart';
 export 'src/types.dart';
 export 'src/widgets.dart';