+++ lib/src/models.dart
@@ -1,0 +1,5 @@
+library models;
+
+import 'package:koyo/src/widgets.dart';
+
+part 'models/ky_page_extra.dart';