+++ lib/src/widgets.dart
@@ -9,6 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:koyo/koyo.dart';
 
 part 'widgets/koyo.dart';
+part 'widgets/animations/ky_page_transition.dart';
 part 'widgets/buttons/ky_button.dart';
 part 'widgets/misc/ky_empty.dart';
 part 'widgets/layout/ky_column.dart';